remove statement
