+++ dinky-admin/src/main/java/org/dinky/controller/TaskController.java
@@ -19,6 +19,7 @@
 
 package org.dinky.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.dinky.data.annotation.Log;
 import org.dinky.data.dto.TaskDTO;
 import org.dinky.data.dto.TaskRollbackVersionDTO;
@@ -36,6 +37,7 @@ import org.dinky.process.exception.ExcuteException;
 import org.dinky.service.TaskService;
 
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -170,6 +172,22 @@ public class TaskController {
             dataTypeClass = Integer.class)
     public Result<TaskDTO> getOneById(@RequestParam Integer id) {
         return Result.succeed(taskService.getTaskInfoById(id));
+    }
+
+
+    @PostMapping("/getPrintTables")
+    @ApiOperation("Get Print Tables")
+    @SuppressWarnings("unchecked")
+    @ApiImplicitParam(name = "statement", value = "Statement", dataType = "String", paramType = "body", required = true)
+    public Result<List<String>> getPrintTables(@RequestBody String statement) {
+        ObjectMapper objectMapper = new ObjectMapper();
+        try {
+            Map<String, String> data = objectMapper.readValue(statement, Map.class);
+            String ss = data.get("statement");
+            return Result.succeed(taskService.getPrintTables(ss));
+        } catch (Exception e) {
+            return Result.failed(e.getMessage());
+        }
     }
 
     @GetMapping(value = "/listFlinkSQLEnv")
