--- dinky-admin/src/main/java/org/dinky/data/model/Statement.java
@@ -1,56 +1,0 @@
-/*
- *
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package org.dinky.data.model;
-
-import java.io.Serializable;
-
-import com.baomidou.mybatisplus.annotation.TableName;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-/**
- * Statement
- *
- * @since 2021/5/28 13:39
- */
-@Data
-@EqualsAndHashCode(callSuper = false)
-@TableName("dinky_task_statement")
-@ApiModel(value = "Statement", description = "Statement Information")
-public class Statement implements Serializable {
-
-    private static final long serialVersionUID = 1646348574144815792L;
-
-    @ApiModelProperty(value = "ID", dataType = "Integer", example = "1", notes = "Unique identifier for the statement")
-    private Integer id;
-
-    @ApiModelProperty(
-            value = "Tenant ID",
-            dataType = "Integer",
-            example = "1001",
-            notes = "ID of the tenant associated with the statement")
-    private Integer tenantId;
-
-    @ApiModelProperty(value = "SQL Statement", dataType = "String", notes = "SQL statement")
-    private String statement;
-}
