+++ dinky-admin/src/main/java/org/dinky/service/impl/TaskServiceImpl.java
@@ -53,6 +53,7 @@ import org.dinky.data.model.UDFTemplate;
 import org.dinky.data.result.Result;
 import org.dinky.data.result.ResultPool;
 import org.dinky.data.result.SqlExplainResult;
+import org.dinky.explainer.printTable.PrintStatementExplainer;
 import org.dinky.function.compiler.CustomStringJavaCompiler;
 import org.dinky.function.pool.UdfCodePool;
 import org.dinky.function.util.UDFUtil;
@@ -68,6 +69,7 @@ import org.dinky.job.JobResult;
 import org.dinky.mapper.TaskMapper;
 import org.dinky.metadata.result.JdbcSelectResult;
 import org.dinky.mybatis.service.impl.SuperServiceImpl;
+import org.dinky.parser.SqlType;
 import org.dinky.process.context.ProcessContextHolder;
 import org.dinky.process.enums.ProcessType;
 import org.dinky.process.exception.ExcuteException;
@@ -81,14 +83,15 @@ import org.dinky.service.FragmentVariableService;
 import org.dinky.service.JarService;
 import org.dinky.service.JobInstanceService;
 import org.dinky.service.SavepointsService;
-import org.dinky.service.StatementService;
 import org.dinky.service.TaskService;
 import org.dinky.service.TaskVersionService;
 import org.dinky.service.UDFTemplateService;
 import org.dinky.service.UserService;
+import org.dinky.trans.Operations;
 import org.dinky.utils.FragmentVariableUtils;
 import org.dinky.utils.JsonUtils;
 import org.dinky.utils.RunTimeUtil;
+import org.dinky.utils.SqlUtil;
 import org.dinky.utils.UDFUtils;
 
 import org.apache.commons.collections4.CollectionUtils;
@@ -99,6 +102,7 @@ import java.io.InputStreamReader;
 import java.io.Reader;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -138,7 +142,6 @@ import lombok.extern.slf4j.Slf4j;
 public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implements TaskService {
 
     private final SavepointsService savepointsService;
-    private final StatementService statementService;
     private final ClusterInstanceService clusterInstanceService;
     private final ClusterConfigurationService clusterCfgService;
     private final JarService jarService;
@@ -546,7 +549,6 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
         return tasks.stream()
                 .peek(task -> {
                     Assert.notNull(task, Status.TASK_NOT_EXIST.getMessage());
-                    task.setStatement(statementService.getById(task.getId()).getStatement());
                 })
                 .collect(Collectors.toList());
     }
@@ -811,5 +813,18 @@ public class TaskServiceImpl extends SuperServiceImpl<TaskMapper, Task> implemen
             treeNodes.add(new TreeNode<>(catalogue.getId(), catalogue.getParentId(), catalogue.getName(), i + 1));
         }
         return treeNodes;
+    }
+
+
+    @Override
+    public List<String> getPrintTables(String statement) {
+        // TODO: 2023/4/7 this function not support variable sql, because, JobManager and executor
+        // couple function
+        //  and status and task execute.
+        final String[] statements = SqlUtil.getStatements(SqlUtil.removeNote(statement));
+        return Arrays.stream(statements)
+                .filter(t -> SqlType.PRINT.equals(Operations.getOperationType(t)))
+                .flatMap(t -> Arrays.stream(PrintStatementExplainer.splitTableNames(t)))
+                .collect(Collectors.toList());
     }
 }
